Adding marquee to player page, getting rid of font scaling
Add documentation

Add breakpoint spacing to player card, player page to ensure consistent styling at all screen sizes and metadata lengths

Update CHANGELOG

Smooth out breakpoints on player page

Smooth playercard breakpoints

Update Changelog
