+++ web/src/components/CustomMarquee/CustomMarquee.scss
@@ -1,3 +1,3 @@
-.marquee-text{
+.marquee-text {
   margin-left: 5px !important; // Prevents marquee text from stopping in the improper location occasionally
 }
