+++ web/src/components/SongInfo/SongInfo.jsx
@@ -1,24 +1,38 @@
 import React from "react";
 import "./SongInfo.scss";
 import { useStatusStore } from "@/App.jsx";
+import CustomMarquee from "../CustomMarquee/CustomMarquee";
 
 import PropTypes from "prop-types";
 
 const SongInfo = ({
     sourceId,
+    ...props
 }) => {
     const info = useStatusStore((state) => state.status.sources[sourceId].info);
+    const artistRef = React.useRef(null);
+    const albumRef = React.useRef(null);
+    const trackRef = React.useRef(null);
 
     return (
-        <div className="song-info">
-            <div className={"artist-name player-info-title"}>{info.artist}</div>
-            <div className={"album-name player-info-album"}>{info.album}</div>
-            <div className={"track-name player-info-track"}>{info.track}</div>
+        <div className="song-info" {...props}>
+            <div ref={artistRef} className={"artist-name"}>
+                <CustomMarquee children={info.artist} containerRef={artistRef} />
+            </div>
+
+            <div ref={albumRef} className={"album-name"}>
+                <CustomMarquee children={info.album} containerRef={albumRef} />
+            </div>
+
+            <div ref={trackRef} className={"track-name"}>
+                <CustomMarquee children={info.track} containerRef={trackRef} />
+            </div>
         </div>
     );
 };
 SongInfo.propTypes = {
     sourceId: PropTypes.any.isRequired,
+    props: PropTypes.any,
 };
 
 export default SongInfo;
